% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosc-package.R
\docType{package}
\name{bosc-package}
\alias{bosc}
\alias{bosc-package}
\title{bosc: Behavioral Oscillation Scores and Phase Analysis}
\description{
R port of the behavioral-oscillations MATLAB toolbox, providing functions
for oscillation scoring, surrogate testing, phase analyses, PPC metrics,
clustering, and visualization for behavioral and neural time series.
}
\references{
Ter Wal, M., Linde-Domingo, J., Lifanov, J., Roux, F., Kolibius, L.,
Gollwitzer, S., Lang, J., Hamer, H., Rollings, D., Sawlani, V.,
Chelvarajah, R., Staresina, B., Hanslmayr, S., & Wimber, M. (2021).
Theta rhythmicity governs the timing of behavioural and hippocampal
responses in humans specifically during memory-dependent tasks.
\emph{Nature Communications}, 12, 7048.
\doi{10.1038/s41467-021-25959-7}

Original MATLAB toolbox: \url{https://github.com/marijeterwal/behavioral-oscillations}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/bosc}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
