% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{open_hdf5}
\alias{open_hdf5}
\title{Load HDF5 files (wrapper)}
\usage{
open_hdf5(path, mode = "r")
}
\arguments{
\item{path}{file path to HDF5.}

\item{mode}{file mode (default "r").}
}
\value{
H5File object.
}
\description{
Thin wrapper around hdf5r::H5File to keep dependency optional.
}
