% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{pairwise_phase_consistency}
\alias{pairwise_phase_consistency}
\title{Pairwise Phase Consistency (1D)}
\usage{
pairwise_phase_consistency(data, dim = 1)
}
\arguments{
\item{data}{array of angles (radians).}

\item{dim}{integer; dimension along which to compute PPC.}
}
\value{
array of PPC values with the chosen dimension collapsed.
}
\description{
Bias-free PPC for angle data along a chosen dimension.
}
