% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_phase_hist}
\alias{plot_phase_hist}
\title{Plot phase histogram}
\usage{
plot_phase_hist(phases, nbins = 12)
}
\arguments{
\item{phases}{numeric vector of phases (radians).}

\item{nbins}{number of bins.}
}
\value{
ggplot object.
}
\description{
Plot phase histogram
}
